Remind user when UserKnownHostsFile is set to /dev/null
When UserKnownHostsFile is set to /dev/null in /etc/ssh/ssh_config,
'kdumpctl propagate' fails because of failure on checking host key.
So we should remind users in that case so that thay can know where
to start to fix the problem. Also remind users when UserKnownHostsFile
is set to "no" just in case.